Adds PromptResponseFormat
