+++ src/assessment.ts
@@ -7,24 +7,3 @@ export enum AssessmentStatus {
   Rejected = 'REJECTED',
   Approved = 'APPROVED',
 }
-
-/**
- * These are the names of the template variables
- * that can be injected into an HTML template
- */
-export enum TemplateVariableModelName {
-  Attribute = 'attribute',
-  BusinessEntity = 'businessEntity',
-  BusinessEntities = 'businessEntities',
-  DataCategory = 'dataCategory',
-  DataCategories = 'dataCategories',
-  Vendor = 'vendor',
-  Vendors = 'vendors',
-  DataSilo = 'dataSilo',
-  DataSilos = 'dataSilos',
-  ProcessingPurpose = 'processingPurpose',
-  ProcessingPurposes = 'processingPurposes',
-  DataPointProperty = 'dataPointProperty',
-  DataPointObject = 'dataPointObject',
-  DataReport = 'dataReport',
-}
